+++ src/sisl/io/siesta/binaries.py
@@ -10,11 +10,13 @@ from typing import Optional
 
 import numpy as np
 
+from sisl.physics.brillouinzone import MonkhorstPack
+
 try:
     from . import _siesta
 
     has_fortran_module = True
-except ImportError:
+except ImportError as _e:
     has_fortran_module = False
 
 import sisl._array as _a
@@ -134,7 +136,7 @@ def _geometry_align(geom_b, geom_u, cls, method):
             f"{cls.__name__}.{method} has non-equal number of supercells, will copy geometry and use file supercell count."
         )
         geom, is_copy = get_copy(geom, is_copy)
-        geom.set_nsc(geom_b.nsc)
+    geom.set_nsc(geom_b.nsc)
 
     # Now for the difficult part.
     # If there is a mismatch in the number of orbitals we will
@@ -788,24 +790,31 @@ class tsdeSileSiesta(dmSileSiesta):
 class hsxSileSiesta(SileBinSiesta):
     """Hamiltonian and overlap matrix file
 
-    This file does not contain all information regarding the system.
+    Historically there are several HSX file versions output from Siesta.
 
-    To ensure no errors are being raised one should pass a `Geometry` with
-    correct number of atoms and correct number of supercells.
-    The number of orbitals will be updated in the returned matrices geometry.
+    Since Siesta 5.3, the HSX file format is containing all the information present in
+    the TranSiesta specific TSHS file-format, plus more information.
+    As such, the TSHS file format is deprecated and will not be used anymore.
+    Therefore users should adapt their workflows to be done with HSX
+    files.
 
-    >>> hsx = hsxSileSiesta("siesta.HSX")
-    >>> HS = hsx.read_hamiltonian() # may fail
-    >>> HS = hsx.read_hamiltonian(geometry=<>) # should run correctly if above satisfied
+    Basically, one should never work with HSX files from Siesta <= 4, since they are
+    not information complete.
 
-    Users are adviced to use the `tshsSileSiesta` instead since that correctly contains
-    all information.
+    Notes
+    -----
+    At some point we will deprecate the older HSX file versions (<= 4).
     """
 
     @property
     def version(self) -> int:
         """The version of the file"""
         return int(_siesta.read_hsx_version(self.file))
+
+    @property
+    def _stored_dtype(self):
+        is_dp = _siesta.read_hsx_is_dp(self.file)
+        return {True: np.float64, False: np.float32}[is_dp]
 
     def _xij2system(self, xij, geometry=None, **kwargs):
         """Create a new geometry with *correct* nsc and somewhat correct xyz
@@ -1178,7 +1187,7 @@ class hsxSileSiesta(SileBinSiesta):
             m = 0
             orbs = []
             for n, l, zeta in zip(*n_l_zeta):
-                if old_values != (n, l, zeta):
+                if old_values != (n, l, zeta) or m > l:
                     old_values = (n, l, zeta)
                     m = -l
                 orbs.append(AtomicOrbital(n=n, l=l, m=m, zeta=zeta, R=-1.0))
@@ -1260,9 +1269,12 @@ class hsxSileSiesta(SileBinSiesta):
     def _r_lattice_v1(self, **kwargs):
         _, na, _, _, _ = _siesta.read_hsx_sizes(self.file)
 
-        cell, nsc, _, _ = _siesta.read_hsx_geom1(self.file, na)
-
+        cell, nsc, _, _ = _siesta.read_hsx_geom1_2(self.file, na)
+        self._log("lattice v1_2")
         return Lattice(cell.T * _Bohr2Ang, nsc=nsc)
+
+    # Same
+    _r_lattice_v2 = _r_lattice_v1
 
     def read_lattice(self, **kwargs) -> Lattice:
         """Read the lattice from the file
@@ -1270,7 +1282,6 @@ class hsxSileSiesta(SileBinSiesta):
         This will always work on new files Siesta >=5, but only sometimes on older
         versions of the HSX file format.
         """
-        self._log("lattice v1")
         return getattr(self, f"_r_lattice_v{self.version}")(**kwargs)
 
     def _r_geometry_v0(self, **kwargs):
@@ -1293,7 +1304,7 @@ class hsxSileSiesta(SileBinSiesta):
 
     def _r_geometry_v1(self, **kwargs):
         self._log(
-            "geometry v1 %r, %r",
+            f"geometry v{self.version} %r, %r",
             kwargs.get("atoms", None),
             kwargs.get("geometry", None),
         )
@@ -1303,10 +1314,12 @@ class hsxSileSiesta(SileBinSiesta):
         # now read coordinates and cell sizes
         _, na, _, _, _ = _siesta.read_hsx_sizes(self.file)
 
-        cell, nsc, xa, _ = _siesta.read_hsx_geom1(self.file, na)
+        cell, nsc, xa, _ = _siesta.read_hsx_geom1_2(self.file, na)
 
         lattice = Lattice(cell.T * _Bohr2Ang, nsc=nsc)
         return Geometry(xa.T * _Bohr2Ang, atoms, lattice=lattice)
+
+    _r_geometry_v2 = _r_geometry_v1
 
     def read_geometry(self, **kwargs) -> Geometry:
         """Read the geometry from the file
@@ -1377,7 +1390,7 @@ class hsxSileSiesta(SileBinSiesta):
         H._csr._D[:, spin] = dS[:]
 
         _mat_siesta2sisl(H)
-        H = H.astype(dtype=kwargs.get("dtype"), copy=False)
+        H = H.astype(dtype=kwargs.get("dtype", self._stored_dtype), copy=False)
 
         # Convert the supercells to sisl supercells
         if no_s // no == np.prod(geom.nsc):
@@ -1396,7 +1409,7 @@ class hsxSileSiesta(SileBinSiesta):
 
         spin, _, no, no_s, nnz = _siesta.read_hsx_sizes(self.file)
         self._fortran_check("read_hamiltonian", "could not read Hamiltonian sizes.")
-        ncol, col, dH, dS, isc = _siesta.read_hsx_hsx1(self.file, spin, no, no_s, nnz)
+        ncol, col, dH, dS, isc = _siesta.read_hsx_hsx1_2(self.file, spin, no, no_s, nnz)
         col -= 1
         self._fortran_check("read_hamiltonian", "could not read Hamiltonian.")
 
@@ -1423,7 +1436,7 @@ class hsxSileSiesta(SileBinSiesta):
         H._csr._D[:, spin] = dS[:]
 
         _mat_siesta2sisl(H)
-        H = H.astype(dtype=kwargs.get("dtype"), copy=False)
+        H = H.astype(dtype=kwargs.get("dtype", self._stored_dtype), copy=False)
 
         # Convert the supercells to sisl supercells
         _csr_from_sc_off(H.geometry, isc.T, H._csr)
@@ -1434,8 +1447,10 @@ class hsxSileSiesta(SileBinSiesta):
 
         return H.transpose(spin=False, sort=kwargs.get("sort", True))
 
+    _r_hamiltonian_v2 = _r_hamiltonian_v1
+
     def read_hamiltonian(self, **kwargs) -> Hamiltonian:
-        """Returns the electronic structure from the siesta.TSHS file"""
+        """Returns the electronic structure from the siesta.HSX file"""
         return getattr(self, f"_r_hamiltonian_v{self.version}")(**kwargs)
 
     def _r_overlap_v0(self, **kwargs):
@@ -1456,6 +1471,9 @@ class hsxSileSiesta(SileBinSiesta):
                 "inconsistent with HSX file."
             )
 
+        # Get dtype
+        dtype = kwargs.get("dtype", self._stored_dtype)
+
         # Create the Hamiltonian container
         S = Overlap(geom, nnzpr=1)
 
@@ -1466,7 +1484,7 @@ class hsxSileSiesta(SileBinSiesta):
         S._csr.col = col.astype(np.int32, copy=False)
         S._csr._nnz = len(col)
 
-        S._csr._D = _a.empty([nnz, 1], dtype=dS.dtype)
+        S._csr._D = _a.empty([nnz, 1], dtype=dtype)
         S._csr._D[:, 0] = dS[:]
 
         # Convert the supercells to sisl supercells
@@ -1483,7 +1501,7 @@ class hsxSileSiesta(SileBinSiesta):
         # Now read the sizes used...
         spin, _, no, no_s, nnz = _siesta.read_hsx_sizes(self.file)
         self._fortran_check("read_overlap", "could not read overlap matrix sizes.")
-        ncol, col, dS, isc = _siesta.read_hsx_sx1(self.file, spin, no, no_s, nnz)
+        ncol, col, dS, isc = _siesta.read_hsx_sx1_2(self.file, spin, no, no_s, nnz)
         col -= 1
         self._fortran_check("read_overlap", "could not read overlap matrix.")
 
@@ -1493,6 +1511,9 @@ class hsxSileSiesta(SileBinSiesta):
                 "passed geometry as the number of atoms or orbitals is "
                 "inconsistent with HSX file."
             )
+
+        # Get dtype
+        dtype = kwargs.get("dtype", self._stored_dtype)
 
         # Create the Hamiltonian container
         S = Overlap(geom, nnzpr=1)
@@ -1504,7 +1525,7 @@ class hsxSileSiesta(SileBinSiesta):
         S._csr.col = col.astype(np.int32, copy=False)
         S._csr._nnz = len(col)
 
-        S._csr._D = _a.empty([nnz, 1], dtype=dS.dtype)
+        S._csr._D = _a.empty([nnz, 1], dtype=dtype)
         S._csr._D[:, 0] = dS[:]
 
         _csr_from_sc_off(S.geometry, isc.T, S._csr)
@@ -1512,9 +1533,154 @@ class hsxSileSiesta(SileBinSiesta):
         # not really necessary with Hermitian transposing, but for consistency
         return S.transpose(sort=kwargs.get("sort", True))
 
+    _r_overlap_v2 = _r_overlap_v1
+
     def read_overlap(self, **kwargs) -> Overlap:
-        """Returns the electronic structure from the siesta.TSHS file"""
+        """Returns the electronic structure from the siesta.HSX file"""
         return getattr(self, f"_r_overlap_v{self.version}")(**kwargs)
+
+    def write_hamiltonian(self, H, **kwargs):
+        """Writes the Hamiltonian to a siesta.HSX file"""
+        # we sort below, so no need to do it here
+        # see hsxSileSiesta.read_overlap for .transpose()
+        H = H.transpose(spin=False, sort=False)
+        if H._csr.nnz == 0:
+            raise SileError(
+                f"{self!r}.write_hamiltonian cannot write "
+                "a zero element sparse matrix!"
+            )
+
+        # Default to writing version 2
+        version = kwargs.get("version", 2)
+        if version not in (1, 2):
+            raise NotImplementedError(
+                f"{self!r}.write_hamiltonian only " "supports one of versions [1, 2]."
+            )
+        sort = kwargs.get("sort", True)
+        bz = kwargs.get("bz", None)  # In case the BZ is an argument
+        dtype = kwargs.get("dtype", np.float64)
+        if dtype not in (np.float32, np.float64):
+            raise ValueError(
+                f"{self!r}.write_hamiltonian only "
+                "supports float32 or float64 in the HSX file format."
+            )
+
+        # Convert to siesta CSR
+        _csr_to_siesta(H.geometry, H._csr, diag=True)
+        H.finalize(sort=sort)
+
+        # we have the required dtype above.
+        # However, the fortran interface requires a double, so we have to
+        # use a bool to signal the stored data.
+        H = H.astype(dtype=np.float64, copy=False)
+        _mat_sisl2siesta(H)
+        csr = H._csr
+
+        # Extract the data to pass to the fortran routine
+        cell = H.geometry.cell
+        xyz = H.geometry.xyz
+
+        # Get H and S
+        if H.orthogonal:
+            self._log("writing orthogonal hamiltonian")
+            s = csr.copy(dims=0)
+            s.empty(keep_nnz=True)
+            s += s.diags(1)
+            missing_diags = s.nnz - csr.nnz
+            if missing_diags:
+                # This should never happen as _csr_to_siesta should ensure
+                # that the diagonal entries always exists.
+                # Hence it gets changed before writing.
+                # Not compeletely optimal, but for now this is OK
+                raise SislError(
+                    f"{self.__class__.__name__}.write_hamiltonian "
+                    "The diagonal elements of your orthogonal Hamiltonian "
+                    f"have not been defined. Got {len(csr) - missing_diags} elements, expected {len(csr)}."
+                )
+            h = csr._D
+            s = s._D
+        else:
+            self._log("writing non-orthogonal hamiltonian")
+            h = csr._D[:, : H.S_idx]
+            s = csr._D[:, H.S_idx]
+
+        # Get shorter variants
+        nsc = H.geometry.nsc[:].astype(np.int32)
+        isc = _siesta.siesta_sc_off(*nsc)
+
+        # Correct arguments depending on BZ
+        kcell = _a.zerosi([3, 3])
+        kdispl = _a.zerosd([3])
+        if isinstance(bz, MonkhorstPack):
+            kcell = np.diag(bz._diag)
+            kdispl = bz._displ
+        else:
+            # TODO, perhaps do something based on the type of BZ
+            for i in range(3):
+                if nsc[i] > 1:
+                    kcell[i, i] = 2
+                else:
+                    kcell[i, i] = 1
+
+        geom = H.geometry
+        atoms = geom.atoms
+
+        # Populate arguments
+        args = []
+
+        def func(atom_):
+            atom, _ = atom_
+            return f"{atom.tag:20s}"[:20], atom.Z, len(atom)
+
+        args.extend(list(map(list, zip(*map(func, atoms.iter(species=True))))))
+
+        # Next values
+        no_max = (atoms.lasto - atoms.firsto[:-1]).max() + 1
+        # Atoms require 'n' and 'zeta' to be >=1
+        n = _a.zerosi([no_max, atoms.nspecies], order="F") + 1
+        l = np.zeros_like(n)
+        zeta = np.zeros_like(n) + 1
+        try:
+            for i_s, (atom, _) in enumerate(atoms.iter(species=True)):
+                for io, orb in enumerate(atom):
+                    n[io, i_s] = orb.n
+                    l[io, i_s] = orb.l
+                    zeta[io, i_s] = orb.zeta
+        except Exception:
+            pass  # should we warn?
+
+        args.extend([n, l, zeta])
+
+        args.extend(
+            [
+                _toF(xyz.T, np.float64, 1 / _Bohr2Ang),
+                atoms.species + 1,
+                atoms.firsto[:-1],
+            ]
+        )
+
+        # Now cell etc.
+        args.append(_toF(cell.T, np.float64, 1 / _Bohr2Ang))
+        args.append(nsc)
+        args.append(isc)  # already in correct shape!
+
+        # Now the matrices
+        args.extend([csr.ncol, csr.col + 1])
+        args.append(_toF(h, np.float64, _eV2Ry))
+        args.append(_toF(s, np.float64))
+        args.append(dtype == np.float64)
+        args.extend([0.0, 0.0, 0.0])
+
+        if version == 2:
+            args.append(kcell)
+            args.append(kdispl)
+
+        # see hsxSileSiesta.read_lattice for .T
+        write_hsx = getattr(_siesta, f"write_hsx{version}")
+        write_hsx(self.file, *args)
+        self._fortran_check(
+            "write_hamiltonian", "could not write Hamiltonian and overlap matrix."
+        )
 
 
 @set_module("sisl.io.siesta")